Move conversion functions to separate file
